feat(overpass): do not download pbf if present

+++ roles/overpass-api/templates/overpass-setup.py
@@ -79,7 +79,9 @@ log.info('download pbf')
 pbf_url = '{{ overpass_pbf_root_url }}/{{ overpass_pbf_region }}-{}.osm.pbf'.format(pbf_datetime.strftime('%y%m%d'))
 pbf_dest = '{{ overpass_database_dir }}/{{ overpass_pbf_region }}-{}.osm.pbf'.format(pbf_datetime.strftime('%y%m%d'))
 
-urlretrieve(pbf_url, pbf_dest)
+if not os.path.exists(pbf_dest):
+    urlretrieve(pbf_url, pbf_dest)
+
 resp = requests.get(pbf_url + '.md5')
 md5_osm = resp.text.split()[0]
 if md5(pbf_dest) != md5_osm:
